Text-to-Image-generator/backend/requirements.txt
Flask
Flask-Cors
Flask-Limiter
python-dotenv
transformers
torch
diffusers
accelerate
safetensors
Pillow
torchvision
